move dashboard api client to stable namespace
