+++ lib/stability_sdk.rb
@@ -1,7 +1,7 @@
 require "stability_sdk/version"
 require "stability_sdk/client"
 require "stability_sdk/cli"
-require "stability_sdk/unstable/dashboard_client"
+require "stability_sdk/dashboard_client"
 
 module StabilitySDK
   class Error < StandardError; end
